GDB 7.5.1
This includes two patches to BFD related to new Mach-O load commands
added in 10.8 (and possibly 10.7 depending on what version of the
devtools you have installed). The first patch is from upstream and will
be released as part of GDB 7.6. The second patch fixes a bug in the
upstream patch to make GDB properly read a binary's start address.  The
second patch is not yet upstream because I don't have a copyright
assignment agreement with the FSF yet, but it has been approved and some
version of it should find itself in GDB 7.6 as well.  You can find the
related email thread here:
http://sourceware.org/ml/binutils/2012-11/msg00320.html